tryout(view)

+++ src/cursor.zig
@@ -4,42 +4,42 @@ pub inline fn print(buf: []u8, comptime fmt: []const u8, args: anytype) []const 
     return std.fmt.bufPrint(buf, fmt, args) catch unreachable;
 }
 
-pub inline fn to(x: u8, y: u8) []const u8 {
+pub inline fn to(x: u16, y: u16) []const u8 {
     var buf: [16]u8 = undefined;
     return print(&buf, "\x1b[{d};{d}H", .{ y, x });
 }
 
-pub inline fn up(n: u8) []const u8 {
+pub inline fn up(n: u16) []const u8 {
     var buf: [16]u8 = undefined;
     return print(&buf, "\x1b[{d}A", .{n});
 }
 
-pub inline fn down(n: u8) []const u8 {
+pub inline fn down(n: u16) []const u8 {
     var buf: [16]u8 = undefined;
     return print(&buf, "\x1b[{d}B", .{n});
 }
 
-pub inline fn right(n: u8) []const u8 {
+pub inline fn right(n: u16) []const u8 {
     var buf: [16]u8 = undefined;
     return print(&buf, "\x1b[{d}C", .{n});
 }
 
-pub inline fn left(n: u8) []const u8 {
+pub inline fn left(n: u16) []const u8 {
     var buf: [16]u8 = undefined;
     return print(&buf, "\x1b[{d}D", .{n});
 }
 
-pub inline fn next_line(n: u8) []const u8 {
+pub inline fn next_line(n: u16) []const u8 {
     var buf: [16]u8 = undefined;
     return print(&buf, "\x1b[{d}E", .{n});
 }
 
-pub inline fn prev_line(n: u8) []const u8 {
+pub inline fn prev_line(n: u16) []const u8 {
     var buf: [16]u8 = undefined;
     return print(&buf, "\x1b[{d}F", .{n});
 }
 
-pub inline fn column(n: u8) []const u8 {
+pub inline fn column(n: u16) []const u8 {
     var buf: [16]u8 = undefined;
     return print(&buf, "\x1b[{d}G", .{n});
 }
